requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
constantly==15.1.0
cryptography==35.0.0
cymysql==0.9.18
daphne==3.0.2
dj-database-url==0.5.0
Django==2.2.10
django-appconf==1.0.4
django-compressor==2.4.1
django-compressor-toolkit==0.6.0
django-session-timeout==0.1.0
fpdf==1.7.2
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.2
mysqlclient==2.0.3
pkg_resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pymodbus==2.5.2
pyOpenSSL==21.0.0
pyserial==3.5
python-decouple==3.4
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
RPi.GPIO==0.7.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
Unipath==1.1
websocket-client==1.2.1
whitenoise==5.3.0
zope.interface==5.4.0
